Remove meta-selinux layer / submodule.
The SELinux stuff was initially used when playing around with this
project.  Focus has shifted to the example 'measured' images and
away from SELinux in an attempt to work from "the ground up".
Removing the meta-selinux submodule prevents confusion as to current
focus.  It will be added later when we're working on securing the
rootfs.