fetch.sh: Restructure LAYERS and fetch.sh accordingly.
The goal here is to make the LAYERS file capable of being a sort of
manifest that can be used to reproduce builds. This means allowing
for the use of an object hash as the desired state of a repo. Next
step is generating a LAYERS file from the current state of the build
directory.