Clarify meaning of 'layers' in JSON.
If layers is omitted from the JSON, is None / null or an empty list
it is assumed that we're not using any meta layers from the repo.
Further when we produce a LAYERS.json file if the root of the repo is
a meta layer we explicitly provide a "./" in the layers list. Previously
we would produce an empty string which caused confusion.